# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Self-hosted mail platform — REST API for the admin console and webmail.",
        "title": "mailez API",
        "contact": {},
        "version": "0.1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/domains": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "domains"
                ],
                "summary": "List domains",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/mailez_backend_internal_core_models.Domain"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/mailez_backend_internal_core_models.APIError"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "domains"
                ],
                "summary": "Create domain",
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/mailez_backend_internal_core_models.Domain"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/mailez_backend_internal_core_models.APIError"
                        }
                    }
                }
            }
        },
        "/domains/{name}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "domains"
                ],
                "summary": "Get domain",
                "parameters": [
                    {
                        "type": "string",
                        "description": "domain name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mailez_backend_internal_core_models.Domain"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "domains"
                ],
                "summary": "Update domain",
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/mailez_backend_internal_core_models.APIError"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "domains"
                ],
                "summary": "Delete domain",
                "parameters": [
                    {
                        "type": "string",
                        "description": "domain name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/mailez_backend_internal_core_models.APIError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "mailez_backend_internal_core_models.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "rate_limited"
                },
                "error": {
                    "type": "string",
                    "example": "something went wrong"
                }
            }
        },
...
        "mailez_backend_internal_core_models.Domain": {
            "type": "object",
            "properties": {
                "anonmail_enabled": {
                    "type": "boolean"
                },
                "comment": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "max_aliases": {
                    "type": "integer"
                },
                "max_quota_bytes": {
                    "type": "integer"
                },
                "max_users": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "signup_enabled": {
                    "type": "boolean"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "SessionCookie": {
            "type": "apiKey",
            "name": "mailez_session",
            "in": "cookie"
        }
    }
}